+++ backend/src/dto/hall.rs
@@ -10,10 +10,10 @@ use uuid::Uuid;
 use crate::error::AppError;
 
 impl HallPayload {
-    pub async fn all(db: &Database, limit: usize) -> Result<Vec<Self>, AppError> {
+    pub async fn all(db: &Database, limit: u32) -> Result<Vec<Self>, AppError> {
         Ok(db
             .halls()
-            .all(limit)
+            .all(limit as usize)
             .await?
             .into_iter()
             .map(Self::from)
