+++ backend/src/dto/location.rs
@@ -47,10 +47,10 @@ pub struct LocationPostPayload {
 }
 
 impl LocationPayload {
-    pub async fn all(db: &Database, limit: usize) -> Result<Vec<Self>, AppError> {
+    pub async fn all(db: &Database, limit: u32) -> Result<Vec<Self>, AppError> {
         Ok(db
             .locations()
-            .all(limit)
+            .all(limit as usize)
             .await?
             .into_iter()
             .map(Self::from)
